requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
av==11.0.0
beautifulsoup4==4.12.2
blessings==1.7
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
Flask==2.3.2
frozenlist==1.3.3
ftransc==7.0.3
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.184
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pafy==0.5.5
pandas==2.0.2
plumbum==1.8.2
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydub==0.25.1
PyJWT==2.8.0
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.31.0
reverso-context-api==0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stripe==5.4.0
telebot==0.0.5
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.16
websocket-client==1.1.0
Werkzeug==2.3.4
yarl==1.9.2
youtube-dl==2021.12.17